refactor(locale)!: use singular locale definition keys (#2932)
