+++ docs/guide/upgrading.md
@@ -495,13 +495,14 @@ Renamed deprecated locale aliases `cz`, `en_IND`, `ge` and removed `global`.
 
 ### Renamed Locale Definitions
 
-The following locale definitions have been renamed
+The following locale definitions have been adjusted to align with Faker's locale definition naming standard:
 
 | old                                         | replacement                                  |
 | ------------------------------------------- | -------------------------------------------- |
 | `faker.definitions.science.chemicalElement` | `faker.definitions.science.chemical_element` |
-| `faker.definitions.system.directoryPaths`   | `faker.definitions.system.directory_paths`   |
-| `faker.definitions.system.mimeTypes`        | `faker.definitions.system.mime_types`        |
+| `faker.definitions.system.directoryPaths`   | `faker.definitions.system.directory_path`    |
+| `faker.definitions.system.mimeTypes`        | `faker.definitions.system.mime_type`         |
+| `faker.definitions.lorem.words`             | `faker.definitions.lorem.word`               |
 
 With that now all our locale data use the following naming scheme:
