+++ src/definitions/lorem.ts
@@ -7,5 +7,5 @@ export type LoremDefinition = LocaleEntry<{
   /**
    * Lorem words used to generate dummy texts.
    */
-  words: string[];
+  word: string[];
 }>;
