+++ src/definitions/system.ts
@@ -7,12 +7,12 @@ export type SystemDefinition = LocaleEntry<{
   /**
    * Returns some common file paths.
    */
-  directory_paths: string[];
+  directory_path: string[];
 
   /**
    * The mime type definitions with some additional information.
    */
-  mime_types: { [mimeType: string]: SystemMimeTypeEntryDefinition };
+  mime_type: { [mimeType: string]: SystemMimeTypeEntryDefinition };
 }>;
 
 /**
