+++ src/locales/ar/lorem/word.ts
