+++ src/locales/base/system/index.ts
@@ -3,12 +3,12 @@
  * Run 'pnpm run generate:locales' to update.
  */
 import type { SystemDefinition } from '../../..';
-import directory_paths from './directory_paths';
-import mime_types from './mime_types';
+import directory_path from './directory_path';
+import mime_type from './mime_type';
 
 const system: SystemDefinition = {
-  directory_paths,
-  mime_types,
+  directory_path,
+  mime_type,
 };
 
 export default system;
