+++ src/locales/base/system/mime_type.ts
