+++ src/locales/cs_CZ/lorem/word.ts
