+++ src/locales/de/lorem/word.ts
