+++ src/locales/dv/lorem/word.ts
