+++ src/locales/el/lorem/word.ts
