+++ src/locales/en/lorem/word.ts
