+++ src/locales/en_BORK/lorem/word.ts
