+++ src/locales/es_MX/lorem/word.ts
