+++ src/locales/fa/lorem/word.ts
