+++ src/locales/fr/lorem/word.ts
