+++ src/locales/fr_CH/lorem/word.ts
