+++ src/locales/he/lorem/word.ts
