+++ src/locales/hy/lorem/word.ts
