+++ src/locales/ja/lorem/word.ts
