+++ src/locales/ko/lorem/word.ts
