+++ src/locales/lv/lorem/word.ts
