+++ src/locales/nl/lorem/word.ts
