+++ src/locales/pl/lorem/word.ts
