+++ src/locales/pt_BR/lorem/word.ts
