+++ src/locales/ru/lorem/word.ts
