+++ src/locales/sk/lorem/word.ts
