+++ src/locales/tr/lorem/word.ts
