+++ src/locales/uz_UZ_latin/lorem/word.ts
