+++ src/locales/vi/lorem/word.ts
